Added advanced products list with pagination, sorting and filtering.

+++ src/services/product.js
@@ -1,4 +1,5 @@
 const ProductRepository = require('../repositories/product')
+const RequestHandlingError = require('../helpers/error')
 
 class ProductService {
 
@@ -9,20 +10,32 @@ class ProductService {
 
     static async getProduct(id) {
         const product = ProductRepository.getProduct(id) 
+        if(!product) throw new RequestHandlingError(404, "Product not found!")
         return product
+    }
+
+    static async getListOfProducts(options) {
+        const products = ProductRepository.getListOfProducts(options)
+        if(!products) throw new RequestHandlingError(404, "Product not found!")
+        return products
     }
 
     static async getAllProducts() {
         const products = ProductRepository.getAllProducts()
+        if(!products) throw new RequestHandlingError(404, "Product not found!")
         return products
     }
 
     static async editProduct(id, productObject) {
+        const product = ProductRepository.getProduct(id) 
+        if(!product) throw new RequestHandlingError(404, "Product not found!")
         const newProduct = ProductRepository.editProduct(id, productObject)
         return newProduct
     }
 
     static async deleteProduct(id) {
+        const product = ProductRepository.getProduct(id) 
+        if(!product) throw new RequestHandlingError(404, "Product not found!")
         const deletedProduct = ProductRepository.deleteProduct(id)
         return deletedProduct
     }
